+++ src/app/pluginLifecycle.ts
@@ -1,7 +1,8 @@
 import { PluginLogger } from '@pivi/agent/logging/pluginLogger';
+import type { App, Plugin } from "obsidian";
 
+import type { PiviApplicationFacades } from "@/app/hostContracts";
 import { registerSelectionToolbarUi } from "@/app/ui/selectionToolbar/SelectionToolbarSurfaceController";
-import type PiviPlugin from "@/main"
 
 import { registerPiviCommands } from "./commandRegistration";
 import { registerEditorSelectionToolbar } from "./editorSelectionToolbarRegistration";
@@ -13,34 +14,38 @@ import { registerPiviViews } from "./viewRegistration";
 
 const logger = new PluginLogger('PluginLifecycle');
 
-export async function initializePiviPlugin(plugin: PiviPlugin): Promise<void> {
-  await measureStartupPhase('settings', () => plugin.loadSettings());
-  registerPiviViews(plugin);
-  registerPiviCommands(plugin);
-  registerPiviSettings(plugin);
+export async function initializePiviPlugin(
+  plugin: Plugin,
+  facades: PiviApplicationFacades,
+  loadSettings: () => Promise<void>,
+): Promise<void> {
+  await measureStartupPhase('settings', loadSettings);
+  registerPiviViews(plugin, facades.chat, facades.sessions, facades.workspace);
+  registerPiviCommands(plugin, facades.chat);
+  registerPiviSettings(plugin, facades.settings, facades.workspace);
   registerEditorSelectionToolbar(plugin, {
     isToolbarEnabled: () => (
-      plugin.settings.editorSelectionToolbar?.enabled !== false
-      && plugin.settings.editorSelectionToolbar.shortcuts.some(item => item.enabled)
+      facades.integrations.settings.editorSelectionToolbar?.enabled !== false
+      && facades.integrations.settings.editorSelectionToolbar.shortcuts.some(item => item.enabled)
     ),
     shouldYieldToNoteToolbar: () => isNoteToolbarTextToolbarActive(plugin.app),
   });
-  registerSelectionToolbarUi(plugin);
+  registerSelectionToolbarUi(facades.integrations, cleanup => plugin.register(cleanup));
 
   plugin.app.workspace.onLayoutReady(() => {
-    void plugin.ensureWorkspaceServices().catch((error: unknown) => {
+    void facades.workspace.ensureWorkspaceServices().catch((error: unknown) => {
       logger.error('Failed to initialize workspace services', error);
     });
   });
 }
 
 export async function persistOpenTabStates(
-  plugin: PiviPlugin,
+  app: App,
 ): Promise<void> {
   // Ensures state is saved even if Obsidian quits without calling onClose().
   const persistOperations: Promise<void>[] = [];
   const errors: unknown[] = [];
-  for (const view of findAllPiviViews(plugin.app)) {
+  for (const view of findAllPiviViews(app)) {
     try {
       const operation = view.getChatHandle()?.maintenance.persistState();
       if (operation) {
