+++ src/app/ui/chatUiCompositionHost.ts
@@ -1,45 +1,5 @@
-import type { CapabilityApprovalPort } from '@pivi/agent/ports';
-import type { OpenSessionState, SessionSummary } from '@pivi/agent/runtime';
-import type { AuxQueryRunner } from '@pivi/agent/runtime/auxQueryRunner';
-import type { PiChatService } from '@pivi/agent/runtime/piChatService';
-import type { SessionMessagePage } from '@pivi/agent/session';
-import type { PiviManagementApprovalPort } from '@pivi/agent/tools/piviManagement';
-import type { ChatPerfRecorder } from '@pivi/pivi-react/store';
-
-import type { PiviChatCompositionHost } from '@/app/hostContracts';
+import type { ChatFacade, SessionsFacade } from '@/app/hostContracts';
 
 /** Composition-only plugin capabilities adapted into core-owned chat ports. */
-export type ChatUiCompositionHost = PiviChatCompositionHost & {
-  getChatPerfRecorder(): ChatPerfRecorder;
-  createChatService(options?: {
-    capabilityApproval?: CapabilityApprovalPort | null;
-    piviManagementApproval?: PiviManagementApprovalPort | null;
-  }): PiChatService;
-  createAuxQueryRunner(): AuxQueryRunner;
-  getSessionList(): SessionSummary[];
-  getOpenSessionSync(id: string): OpenSessionState | null;
-  getOpenSessionById(id: string): Promise<OpenSessionState | null>;
-  openRecentSessionMessages(id: string, limit: number): Promise<SessionMessagePage | null>;
-  readOlderSessionMessages(
-    id: string,
-    beforeEntryId: string,
-    limit: number,
-  ): Promise<SessionMessagePage | null>;
-  createOpenSession(options?: {
-    sessionId?: string;
-    sessionFile?: string;
-  }): Promise<OpenSessionState>;
-  openSessionByFile(sessionFile: string): Promise<OpenSessionState>;
-  deleteSession(id: string): Promise<void>;
-  deleteSessionFile(sessionFile: string, id?: string | null): Promise<void>;
-  renameSession(
-    id: string,
-    title: string,
-    titleSource?: OpenSessionState['titleSource'],
-  ): Promise<void>;
-  updateSession(id: string, updates: Partial<OpenSessionState>): Promise<void>;
-  forkSessionAt(
-    openSession: OpenSessionState,
-    atEntryId: string,
-  ): Promise<{ sessionFile: string; sessionId: string } | null>;
-};
+export type ChatUiCompositionHost = ChatFacade;
+export type ChatUiSessionHost = SessionsFacade;
